app/requirements.txt
chainlit==2.8.3
openai==2.3.0
python-dotenv==1.1.1
httpx==0.28.1
pydantic==2.12.2
typing-extensions>=4.0.0
elasticsearch==9.1.1
mcp
elastic-opentelemetry
opentelemetry-instrumentation-openai
opentelemetry-instrumentation-elasticsearch
opentelemetry-instrumentation-aiohttp-client
opentelemetry-instrumentation-asyncio
opentelemetry-instrumentation-httpx
